Spiking improved resource abstraction - tidy
